Remove CallNumber normalization
The comparison process is now _much_ simpler. The two call numbers are parsed as floats and compared.
If equal, then their cutters are compared. This led to a pretty huge increase in speed:

the test:

    require "vendor/autoload.php";

    for( $i = 0; $i < 1000000; $i++ ) {
        Dewey::compare("741.1234 A123c", "741.1234 A12", ">=");
    }

and the results:

w/ normalization:

    > time php test.php

    real    0m52.567s
    user    0m45.933s
    sys     0m0.329s

w/o normalization:

    > time php test.php

    real    0m18.844s
    user    0m17.792s
    sys     0m0.081s

So that's pretty rad! _And_ none of the tests broke!